pyproject.toml
[project]
dependencies = [
  "strawberry-graphql>=0.129.0",
  "uvicorn>=0.18.3",
  "prisma>=0.6.6",
  "fastapi>=0.82.0",
  "pydantic>=1.10.2",
]
requires-python = ">=3.7"
[project.optional-dependencies]

[tool.pdm]
allow_prereleases = true

[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true

[tool.pdm.dev-dependencies]
dev = [
  "flake8>=5.0.4",
  "black>=22.8.0",
  "pdbpp>=0.10.3",
  "bpython>=0.23",
  "pytest>=7.1.3",
  "rich>=12.5.1",
  "libcst>=0.4.7",
  "mypy>=0.971",
  "typer>=0.6.1",
  "httpx>=1.0.0b0",
  "jsondiff>=2.0.0",
  "python-dateutil>=2.8.2",
]

[tool.pdm.scripts]
cli = "python cli.py"
server = "uvicorn app:app --reload"

[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517>=1.0.0"]

[tool.black]
extend-exclude = '''
tests/codegen/snapshots/
'''
line-length = 88

[tool.isort]
combine_as_imports = true
combine_star = true
indent = 4
known_django = ["django"]
known_first_party = ["strawberry"]
known_graphql = ["graphql"]
known_pytest = ["pytest"]
lines_after_imports = 2
profile = "black"
sections = ["FUTURE", "STDLIB", "PYTEST", "THIRDPARTY", "DJANGO", "GRAPHQL", "FIRSTPARTY", "LOCALFOLDER"]
src_paths = ["strawberry", "tests", "scripts"]

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]
